Update simple.cgi
I've been toying with Lincoln Stein's original script to make the Web version of Eliza look more like a chatbot.
By default it has been set to use Chatbot::dude which is an Eliza hybrid I am working on, and the doc.txt rules file, which is based on the script from the Big Lebowski that I found at https://imsdb.com/scripts/Big-Lebowski,-The.html.
This script works also with the regular Chatbot::Eliza and her rules test. 
In addition I would like to share the fact that the original Stein script I received with the ChatbotEliza package had a couple of errors that made it malfunction. I have fixed those.